fiddling with parameters

+++ MarketSim/src/Commodity.java
@@ -3,7 +3,7 @@ public class Commodity {
 
 	public int commoditynumber;
 	public String name;
-	public int marketprice = -1;
+	public double marketprice = -1;
 	
 	public Commodity(String n, int cn)
 	{
